+++ jwtex/src/main/java/com/soromiso/jwtex/service/UserService.java
@@ -7,9 +7,9 @@ import com.soromiso.jwtex.domain.user.User;
 public interface UserService {
     public User save(User user);
     public List<User> getAllUsers();
-    public boolean existsById(Long id);
-    public User getUserById(Long id);
-    public User updateUser(Long id, User user);
-    public void deleteUser(Long id);
+    public boolean existsById(Integer id);
+    public User getUserById(Integer id);
+    public User updateUser(Integer id, User user);
+    public void deleteUser(Integer id);
     public User findByEmail(String email);
 }
